+++ src/questionnaire/trackSectors.ts
