+++ src/questionnaire/trackStructureSizes.ts
